Implement HTTP Digest Authentication, RFC 2617
This adds capability to respond to HTTP Digest Authentication challenge;
at the same time, does the following:

* introduce common base class for BasicAuthorization &
  DigestAuthorization, named Authorization

* remove BasicAuthorization's inheritance from peer.Header (this is a BC
  break, but helpers have been added that keep old behaviour in the way
  the class had been intented to be used.

* members have been made non-public (another BC break, but you shouldn't
  access foreign classes members + its not probable that these members
  would be accessed somewhere outside of this package)

With the capabilities in place, you still need to handle HttpConnection
properly to make it talk HTTP digest, ie. it will not automatically
re-try HTTP requests with credentials if responded to with a HTTP 401
status code.